fix trigger CI
